Rewrite and simplify redis connection.
Now that the redis driver is mature, we can drop all the specialized
parsing and late handling of authentication. We can also use in-built
key prefixing.

This also takes it down to a single env var for the URL rather than
one per vendor: if the driver can handle it, they can pass it. The
prefix is also env configurable.